Move part of ObserveTimeStepVolume into ObserveConstantsPerElement
ObserveConstantsPerElement can be used as a base class for any observation
event that will observe data that has a single value per Element.
